rename window-unmaximize to window-restore
