+++ packages/react/src/themes/teams/components/Icon/svg/icons/windowRestore.tsx
